==> 07_volume_damage.R <==


# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# LInk trap data with damage volume and RS observation
# ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

# Goal: do beetle counts/dynamics preds (+ lags) predict tree damage?  

# from LWF: the damage volume+ geometries
#  - read tables
#  - merge geometries
#  explore damage per year
# RS disturbance and forests: 
#  - read RS data from Cornelius: 
# - do amage data correlate with RS beetle estimation? YES!
# beetle counts data:
# - how to link them with beetle trap data?
# - select the respective region per year
# - select in terra oes not work well: convert location to raster, and extract forest_reviewr1 for trap data
# - merge by years with beetle counts and damage
# - do beetle counts fit over forest damage??

# to do:
# read data
# expland teh geometries for teh damage data, to plot the damage with ggplot
# how to extrapolated beetle counts from the point-base data? - create a map and overlay this with the reporte damage data
# use a spruce layer for it
# 
  

rm(list=ls()) 

# Libs --------------------------------------------------------------------------
library(dplyr)
library(sf)
library(raster)
library(terra)
library(data.table)
library(ggplot2)
library(ggpubr)

library(lubridate)
library(fasterize)
library(tidyr)


# Stats
library('purrr')
library("tibble")
library('tidyr')
library("knitr")

library('readr')

library(ggeffects)

library(MASS)
library(car)     # for VIF
library(glmmTMB) # many families
library(segmented)  # for breakpoint analysis 


library(lme4) #pseudo-R2 and and mixed model functionality
library(MuMIn) #dredge function for comparing models, AIC, marginal R^2 for mixed models
library(sjmisc) #pseudo R^2 - but I think gets loaded as a dependency
library(DHARMa) #model diagnostics
library(effects) #what do my marginal effects look like?
library(performance) #binomial model diagnostics
library(emmeans) #post hoc for categorical predictors

# test for autocorrelation
library(lmtest)

# colors
library(RColorBrewer)

library(mgcv)
library(mgcViz)
library(lme4)
library(gratia)
#library(knitr)   # for table outputs
library(stringr) 



# Data --------------------------------------------------------------------------

## Input data ----------------------------------------------------
### read beetle data 
# load cleaned data
load("outData/ips_counts.Rdata")
load("outData/spatial.Rdata")
dat_dynamics <- fread( 'outTable/beetle_dynamic_indicators.csv')


source("my_functions.R")
# Get beetle counts - corrected, instead of previous 'dat'
# - dat.ips.clean      # dat <- fread(paste(myPath, outFolder, "dat.csv", sep = "/"))
# - max.diff.doy       # get DOY of max increase
# - ips.aggreg         # get DOY of max increase
# - ips.year.avg       # sum beetles/year, 
# - df.daily           # avg/betles/year per trap

# Spatial data: 
# - xy_sf_fin  # XY as sf data, filtered to have only one globalid per trap
# - xy_sf_expand  # XY as sf data, one point for trap per every year, 1106 rows



### Tree damage volume & districts shp  ---------------------
sf_districts <-  st_read('rawData/damage_volume/AELF_Revier_20240502/AELF_Revier_bis20240424_LWF_20240502.shp')
df_damage    <-  fread('rawData/damage_volume/Bavaria_TreeDamageData_IpsTypographus_2015-2021.csv', dec = ',')

### disturbance rasters 
disturb_year <- rast('rawData/germany114/disturbance_year_1986-2020_germany.tif')
disturb_type <- rast('rawData/fire_wind_barkbeetle_germany.tif')
forest       <- rast('rawData/germany114/forestcover_germany.tif')

crs(disturb_year) <- "EPSG:3035"
# check data: how many regions I have?
sf_ID   <- unique(sf_districts$forstrev_1)  # 336 districts numbers
df_ID   <- unique(df_damage$AELF_district_ID)  # 337 district numbers


### Trap catches: sums/year/trap --------------------------------------------
# add beetle counts to trap locations per year 
sum_ips <- 
  dat.ips.clean %>% 
  group_by(year,falsto_name) %>% 
  dplyr::summarize(sum_ips = sum(fangmenge, na.rm = T)) 

# convert to sf object, crs 3035
sum_ips_sf <-  xy_sf_expand %>%  left_join(sum_ips, by = c("falsto_name", 'year')) 
#st_is_valid(sum_ips_sf_trans)

# change projection & convert to terra
my_crs = st_crs(disturb_year)

# change projection & convert to terra
# districts = polygons
sf_districts_trans <- st_transform(sf_districts, crs = my_crs) # st_crs(disturb_year))
v_districts        <-  terra::vect(sf_districts_trans) # convert to terra

sum_ips_sf_trans  <- st_transform(sum_ips_sf, crs = my_crs)
v_sum_ips         <- vect(sum_ips_sf_trans)


# works well for terra! 
plot(forest)
plot(disturb_type)
plot(v_sum_ips, add = T)



# simplify terra object:
v_districts_simpl <- simplifyGeom(v_districts, tolerance=20, preserveTopology=TRUE, makeValid=TRUE)
v_sum_ips_simpl   <- simplifyGeom(v_sum_ips, tolerance=20, preserveTopology=TRUE, makeValid=TRUE)


#writeVector(v_sum_ips_simpl, 'outSpatial/XY_beetle_sum_years.gpkg', overwrite=TRUE) 

#  does simplification work? yes!!
plot(v_districts_simpl)
plot(v_sum_ips_simpl, add = T, col = 'red')

identical(crs(v_districts_simpl), crs(v_sum_ips_simpl))


### Tree disturbances rasters ------------------------------------------------------------
# there is an aditiona row in the damage data with district ID 0 - I can remove that

#crop  rasters to Bavaria 
dist_year_crop <- terra::crop(disturb_year, v_districts_simpl)
dist_type_crop <- terra::crop(disturb_type, v_districts_simpl)
forest_crop    <- terra::crop(forest,       v_districts_simpl)

# process spatial data: check geometries, change projection to Corneliuses data, convert to raster the AELF districts
crs(disturb_type) <- crs(dist_year_crop)
crs(forest_crop)  <- crs(dist_year_crop)


# Process department shp and extract revier ID 
sf_simpled <- st_as_sf(v_districts_simpl) #st_simplify(sf_districts_trans, dTolerance = 20, preserveTopology = TRUE)

identical(crs(sf_simpled), crs(v_districts_simpl))

# get extend and set correct projectipn
ext <- as(extent(sf_simpled), 'SpatialPolygons')  # treat raster as from raster, not terra package
ext <- st_as_sf(ext)

identical(crs(sum_ips_sf_trans), crs(ext))

# efine teh crs first - it was NA
st_crs(ext) <- my_crs# st_crs(raster(disturb_year))

#ext  <- st_transform(ext, crs = my_crs)

st_crs(sf_simpled) == st_crs(ext) 
st_crs(disturb_year) == st_crs(ext) 
st_crs(sum_ips_sf_trans) == st_crs(ext) 

# create a base raster to rasteroize to it
base_raster = raster(dist_year_crop)

# Create grid index for each pixel
grid_sel     <- st_intersection(sf_simpled, ext)
grid_sel_ras <- fasterize::fasterize(grid_sel,
                                     base_raster,
                                     #raster(dist_year_crop), 
                                     field = "forstrev_1") # name the new rasters as polygon ids  #, field = "forstrev_1"

# assigne the crs manually! differet assigenement for vector and raster
crs(grid_sel_ras)              <- crs(sf_simpled)
st_crs(sum_ips_sf_trans)       <- crs(sf_simpled)

# Extract CRS again after transformation
crs_raster <- crs(sf_simpled)
crs_vector <- st_crs(sum_ips_sf_trans)

# Confirm they are now the same
identical(as.character(crs_raster), crs_vector$wkt)  # Should return TRUE


# check if the crs fits! YES
st_crs(grid_sel_ras)       == st_crs(sf_simpled)
st_crs(grid_sel_ras)       == st_crs(sum_ips_sf_trans)


# extract the 'reviers ID' (from the districts) to the beetle counts data 
district_ID <- terra::extract(rast(grid_sel_ras),
                              v_sum_ips_simpl,
                            #  vect(sum_ips_sf_trans), 
                              df = TRUE, bind = TRUE )


## Processing -----------------------------------------------------
### Link beetle counts vs damage volume per department -------------------------------------------------------
ips_damage_district <- district_ID %>% 
  as.data.frame() %>% 
  dplyr::rename(., forstrev_1 = layer)


# join trap counts with volume damage data and beetle dynamics indicators;
# average per trap pair to not have duplicate records
ips_damage_merge <- 
  ips_damage_district %>% 
  left_join(df_damage, by = c('forstrev_1' = 'AELF_district_ID',
                              'year' = 'Year')) %>% 
  dplyr::select(-c(id, AELF_district_name,AELF_name, AELF_ID,
                 pest_species, tree_species, unit,
                 globalid )) %>% 
  dplyr::rename(trapID = falsto_name) %>% 
   
  # add beetle dynamics variables 
  left_join(dat_dynamics) %>%
  group_by(trapID)  %>% 
  dplyr::arrange(year,.by_group = TRUE) %>% 
   mutate(sum_ips_lag1  = dplyr::lag(sum_ips, n = 1, default = NA),  # 
         sum_ips_lag2  = dplyr::lag(sum_ips, n = 2, default = NA),  # , order_by = year
         agg_doy_lag1  = lag(agg_doy, n = 1, default = NA),
         agg_doy_lag2  = lag(agg_doy, n = 2, default = NA),
         peak_doy_lag1  = lag(peak_doy, n = 1, default = NA),
         peak_doy_lag2  = lag(peak_doy, n = 2, default = NA),
         peak_diff_lag1  = lag(peak_diff, n = 1, default = NA),
         peak_diff_lag2  = lag(peak_diff, n = 2, default = NA),
         lag1_damaged_volume_total  = lag(damaged_volume_total_m3, n = 1, default = NA)
        )  %>%
  
  ungroup(.) %>% 
  mutate(pairID = as.factor(pairID),
         trapID = as.factor(trapID)) #%>% 
    



# Check if lags are correct!!!! YES!!!

ips_damage_merge  %>% 
   dplyr::filter(trapID == "Anzinger_Forst_1") %>% 
     dplyr::select(trapID, year, sum_ips, sum_ips_lag1, sum_ips_lag2,
                   agg_doy, agg_doy_lag2, damaged_volume_total_m3, lag1_damaged_volume_total )
  


# remove NA values
ips_damage_clean <- na.omit(ips_damage_merge)


# Group by pairID and year, then calculate mean for all continuous variables
ips_damage_pairID <- ips_damage_merge %>%
  group_by(pairID, year) %>%
  summarise(across(where(is.numeric), \(x) mean(x, na.rm = TRUE)), .groups = "drop") %>% 
  mutate(across(where(is.numeric), ~ifelse(is.nan(.), NA, .))) 



### Prepare RS data   -------------------------------------------------------------
# merge disturbance rasters together; check n of cells 
ncell(grid_sel_ras)
ncell(dist_year_crop)
ncell(dist_type_crop)
ncell(forest_crop)


# extract values:
ID     = terra::values(grid_sel_ras)
year   = terra::values(dist_year_crop, mat = F)
type   = terra::values(dist_type_crop, mat = F)
forest = terra::values(forest_crop, mat = F)

# merge into one df
df_disturb <- data.frame(ID = ID,
                         year = year,
                         type = type,
                         forest = forest
                         ) %>% 
  dplyr::filter(!is.na(ID))

# What to get: per district: how many cells, forest, disturbances, harvest& bark beetles
df_forest <- df_disturb %>% 
  group_by(ID) %>%
  dplyr::summarize(forest86 = sum(forest == 1, na.rm = TRUE),
                   area_cells = n())

# RS get disturbances and type 2015-2020
df_disturb_sum <- 
  df_disturb %>% 
  dplyr::filter(year %in% 2015:2021) %>% 
  dplyr::select(!forest) %>% 
  dplyr::filter(!is.nan(type)) %>%        # Exclude rows where 'type' is NaN
  group_by(ID, year, type) %>% 
    dplyr::summarize(count = n(), .groups = "drop") %>% 
  pivot_wider(
    names_from = type,          # Create new column names from the 'type' values
    values_from = count,        # Fill these new columns with values from 'count'
    names_prefix = "type_",     # Optional: prefix for new column names for clarity
    values_fill = list(count = 0)  # Fill missing values with 0
  )


# merge forest and disturbances from RS data per ASFL districts
df_RS <- df_disturb_sum %>% 
  full_join(df_forest, by = join_by(ID)) %>% 
  dplyr::rename(RS_wind_beetle = type_1,
                RS_fire    = type_2,
                RS_harvest = type_3) %>% 
  mutate(RS_sum = RS_wind_beetle + RS_fire +RS_harvest) %>% 
  dplyr::select(c(ID,  year,  RS_wind_beetle)) %>% 
  group_by(ID)
  

# type_1 = wind & beetles
# type_2 = fire
# type_3 = harvest



### Merge RS with damage data -----------------------------------------------
df_merged_RS_damage_district <- df_RS %>% 
  full_join(df_damage, by = c("ID"   = "AELF_district_ID",
                              "year" = "Year")) %>% 
  mutate(AELF_name = str_replace_all(AELF_name, "[^A-Za-z0-9\\s]", ""),
         AELF_district_name  = str_replace_all(AELF_district_name , "[^A-Za-z0-9\\s]", ""))  %>% 
  dplyr::select(!c( AELF_ID, tree_species, unit, pest_species ))# %>%  # keep: AELF_district_name, AELF_name,





#### MAP: correlations per XY: beetle counts vs damage volume 
# df_cor_counts_damage <- 
#   ips_damage_clean %>% 
#   #dplyr::filter(!ID %in% c(0, 50104, 60613,62705)) %>% # exlude if there is missing data
#   group_by(trapID,forstrev_1        ) %>% 
#   dplyr::summarize(spearm_cor_lag0 = cor(damaged_volume_total_m3, sum_ips, 
#                                            method = "spearman", use = "complete.obs"),
#                    spearm_cor_lag1 = cor(damaged_volume_total_m3, sum_ips_lag1 , 
#                                             method = "spearman", use = "complete.obs"),
#                    spearm_cor_lag2 = cor(damaged_volume_total_m3, sum_ips_lag2 , 
#                                         method = "spearman", use = "complete.obs"))


#### MAP: correlations per district: beetle counts vs RS damage 
# merge based on the district ID, subset the data as the traps do not cover the whole 
# bavaria's districts

# Final tables ------------------------------------------

# keep datasets separated, as RS has lower number of years
# run analysis for trap vs RS damage
# trap vs damaged volume
df_merged_RS_damage_district_lag1 <- df_merged_RS_damage_district %>% 
  group_by(ID) %>% 
  dplyr::arrange(year, .by_group = TRUE) %>%
  mutate(lag1_RS_wind_beetle  = lag(RS_wind_beetle , n = 1, default = NA)) #%>% 


# check if correct:
df_merged_RS_damage_district_lag1 %>% 
  dplyr::filter(ID == 10504) %>% 
  dplyr::select(ID, year, RS_wind_beetle , lag1_RS_wind_beetle) #%>% # sum_ips, sum_ips_lag1

# average all values on pairID level  ----------------------------------------------------------
# skip also NA values, as I am getting lag 1 and lag2
df_catch_RS_damage_pairID <-  
  ips_damage_clean %>%  
  left_join(df_merged_RS_damage_district_lag1, 
            by = c("forstrev_1" =   "ID",
                   "year"       = "year",
                   "damaged_volume_total_m3" = "damaged_volume_total_m3") ) %>% 
  #mutate(trapID = as.factor(trapID)) %>% 
  group_by(pairID, year) %>% 
  summarise(across(where(is.numeric), ~ mean(.x, na.rm = TRUE)), #.names = "mean_{.col}"),
            .groups = "drop") %>%
   ungroup(.) %>% 
  mutate(log_sum_ips = log(sum_ips)) %>% 
  mutate(f_year = as.factor(year))

# have only 1 lag in my predictors to have longer time series for break point analysis  -------------------------------
df_sub_1lag <- ips_damage_merge %>%  
  left_join(df_merged_RS_damage_district_lag1, 
            by = c("forstrev_1" =   "ID",
                   "year"       = "year",
                   "damaged_volume_total_m3" = "damaged_volume_total_m3") ) %>% 
 # mutate(trapID = as.factor(trapID)) %>% 
  group_by(pairID, year) %>% 
  summarise(across(where(is.numeric), ~ mean(.x, na.rm = TRUE)), #.names = "mean_{.col}"),
            .groups = "drop") %>%
  ungroup(.) %>% 
  mutate(log_sum_ips = log(sum_ips)) %>% 
  mutate(f_year = as.factor(year)) %>% 
  dplyr::select( pairID,   year, sum_ips, sum_ips_lag1, damaged_volume_total_m3,
                 RS_wind_beetle , lag1_RS_wind_beetle,lag1_damaged_volume_total
                 )


summary(df_catch_RS_damage_pairID)

###### create final tables for RS data and volume data 

fin_dat_damage <- df_catch_RS_damage_pairID %>% 
  dplyr::select(-RS_wind_beetle) %>% 
  na.omit() 

dim(fin_dat_damage)


fin_dat_damage %>% 
  ggplot(aes(x = log(sum_ips),
             y = log(damage_vol))) +
  geom_point() +
  geom_smooth()

# do not remove outliers here 
fin_dat_RS <- df_catch_RS_damage_pairID %>% 
  na.omit()

dim(fin_dat_RS)




## Tab for spagetti plot -----------------

df_full_years_to_spagetti_plot <-  
  ips_damage_merge %>%  
  left_join(df_merged_RS_damage_district,
            by = c("forstrev_1" =   "ID",
                   "year"       = "year",
                   "damaged_volume_total_m3" = "damaged_volume_total_m3") )  %>% 
  mutate(trapID = as.factor(trapID)) %>% 
  group_by(pairID, year) %>% 
  dplyr::summarise(sum_ips    = mean(sum_ips, na.rm = T),
            agg_doy    = mean(agg_doy, na.rm = T),
            peak_doy   = mean(peak_doy, na.rm = T),
            peak_diff  = mean(peak_diff, na.rm = T),
            damage_vol = mean(damaged_volume_total_m3, na.rm = T),
            RS_wind_beetle = mean(RS_wind_beetle, na.rm = T)#,
  ) %>%
  ungroup(.) %>% 
  mutate(#log_sum_ips = log(sum_ips),
         RS_wind_beetle_ha = RS_wind_beetle*0.09,  # update values for plotting
         damage_vol_k = damage_vol/1000) #%>% 
#  dplyr::filter(!pairID %in% c('Peiting', "Eschenbach_idOPf"))

# Replace or filter out zero values before plotting
df_full_years_to_spagetti_plot$damage_vol_k <- ifelse(
  df_full_years_to_spagetti_plot$damage_vol_k <= 0.01, 
                                                    0.01, 
                                                    df_full_years_to_spagetti_plot$damage_vol_k)

# Analysis ------------------------------------------------------------------
## Break point analysis ------------------------------------------------------

df_sub_1lag 

# make a unique table to have more data points  f0r ground survey
df_sub_1lag_damage <- df_sub_1lag %>% 
  dplyr::select(-RS_wind_beetle, lag1_RS_wind_beetle) %>% 
  na.omit()


df_sub_1lag_RS <- df_sub_1lag %>% 
  dplyr::select(-damaged_volume_total_m3, lag1_damaged_volume_total) %>% 
  na.omit()

# replace by from  ips_damage_pairID - contains all 2015-2021 values

### Break point: counts vs damage --------------------
plot(damaged_volume_total_m3 ~ sum_ips, df_sub_1lag_damage)

# try on raw data
# Fit a mixed-effects model with sum_ips as a predictor
base_model <- lmer(damaged_volume_total_m3 ~ sum_ips + (1 | pairID), data = df_sub_1lag_damage)

# Apply segmented regression to detect breakpoints
seg_model <- segmented(lm(damaged_volume_total_m3 ~ sum_ips, data = df_sub_1lag_damage), 
                       seg.Z = ~sum_ips, psi = list(sum_ips = median(df_sub_1lag_damage$sum_ips, na.rm = TRUE)))  # Initial guess

# Summarize results
summary(seg_model)

# Extract detected breakpoints
breakpoints <- seg_model$psi
(breakpoints)

## try log transformation for minimise teh variability  

# Log transformation (adding +1 to avoid log(0))
df_sub_1lag_damage$log_sum_ips <- log(df_sub_1lag_damage$sum_ips + 1)
df_sub_1lag_damage$log_damage_volume <- log(df_sub_1lag_damage$damaged_volume_total_m3 + 1)
df_sub_1lag_damage$log_sum_ips_lag1 <- log(df_sub_1lag_damage$sum_ips_lag1 + 1)

# Fit new segmented model: lag0
lmer_model_log <- lmer(log_damage_volume ~ log_sum_ips +  (1 | pairID), data = df_sub_1lag_damage)
seg_model_log_damage <- segmented(lm(log_damage_volume ~ log_sum_ips, data = df_sub_1lag_damage), 
                           seg.Z = ~log_sum_ips, 
                           psi = list(log_sum_ips = median(df_sub_1lag_damage$log_sum_ips, na.rm = TRUE)))

summary(seg_model_log_damage)
# Fit new segmented model: lag1
# lmer_model_log_lag1 <- lmer(log_damage_volume_lag1 ~ log_sum_ips_lag1 +  (1 | pairID), data = df_sub_1lag_damage)
# seg_model_log_lag1 <- segmented(lm(log_damage_volume ~ log_sum_ips_lag1, data = df_sub_1lag_damage), 
#                            seg.Z = ~log_sum_ips_lag1, 
#                            psi = list(log_sum_ips_lag1 = median(df_sub_1lag_damage$log_sum_ips_lag1, na.rm = TRUE)))

confint_breakpoints_damage <- confint(seg_model_log_damage)
confint_breakpoints_damage

# Extract detected breakpoints
breakpoints_damage_log <- seg_model_log_damage$psi
(breakpoints_damage_log)

#breakpoints_log_lag1 <- seg_model_log_lag1$psi
#(breakpoints_log_lag1) # very similar results with predicting the next year value, just larger uinterval- stay with current year

# Plot results
p_break_point_damage <- ggplot(df_sub_1lag_damage, aes(x = log_sum_ips, y = log_damage_volume)) +
  geom_point(alpha = 0.5) +
  geom_line(aes(y = predict(seg_model_log_damage)), color = "red", lwd = 1.2) +
  geom_vline(xintercept = breakpoints_damage_log[, 2], linetype = "dashed", color = "blue") +
  labs(title = "Field Observed",
       x = "Trap Catch (Beetle Density)",
       y = "Tree Mortality (m³)") +
  theme_classic2()

p_break_point_damage

# Count records where sum_ips < threshold
num_below_threshold <- sum(ips_damage_pairID$sum_ips < se_upper_threshold, na.rm = TRUE)

# Display result
num_below_threshold

# Total number of records
total_records <- nrow(ips_damage_pairID)

# Percentage below threshold
percent_below <- (num_below_threshold / total_records) * 100

# Display result
percent_below


### RS ---------------------------------------------------------- 

df <- na.omit(df_sub_1lag_RS)

plot(RS_wind_beetle   ~ sum_ips, df)


#### log transformation for minimise teh variability  

# Log transformation (adding +1 to avoid log(0))
df$log_sum_ips        <- log(df$sum_ips + 1)
df$log_RS_wind_beetle <- log(df$RS_wind_beetle + 1)

# Fit new segmented model
lmer_model_log <- lmer(log_RS_wind_beetle ~ log_sum_ips +  (1 | pairID), data = df)
seg_model_log_RS <- segmented(lm(log_RS_wind_beetle ~ log_sum_ips, data = df), 
                           seg.Z = ~log_sum_ips, 
                           psi = list(log_sum_ips = median(df$log_sum_ips, na.rm = TRUE)))

# Check model fit
summary(seg_model_log_RS)
confint(seg_model_log_RS)
confint_breakpoints_RS <- confint(seg_model_log_RS)

# Extract detected breakpoints
breakpoints_log_RS <- seg_model_log_RS$psi
(breakpoints_log_RS)

# Plot results
p_break_point_RS <- ggplot(df, aes(x = log_sum_ips, y = log_RS_wind_beetle)) +
  geom_point(alpha = 0.5) +
  geom_line(aes(y = predict(seg_model_log_RS)), color = "red", lwd = 1.2) +
  geom_vline(xintercept = breakpoints_log_RS[, 2], linetype = "dashed", color = "blue") +
  labs(title = "RS Observed",
       x = "Trap Catch (Beetle Density)",
       y = "RS Tree Mortality (log m2)") +
  theme_classic2()


# 
ggarrange(p_break_point_damage, p_break_point_RS, ncol = 2, nrow = 1,
          labels = c("[a]", "[b]"))

# Find lags and predictors -------------------------------------------------------
## Find lags: damage volume  -------------
dependent_damage <-  c("damage_vol")

# Initialize a data frame to store AIC values and deviance explained
model_metrics_damage <- data.frame(Predictor = character(), 
                                  Dependent = character(), 
                                  AIC = numeric(), 
                                  R_squared = numeric())

# list predictors to test
selected_predictors <- c('sum_ips', 'sum_ips_lag1','sum_ips_lag2',
                         #'log_sum_ips', 'log_sum_ips_lag1','sum_ips_lag2',
                         'agg_doy'   , 'agg_doy_lag1', 'agg_doy_lag2' ,
                         'peak_doy', 'peak_doy_lag1','peak_doy_lag2',
                         'peak_diff', 'peak_diff_lag1',  'peak_diff_lag2'
) 



# Loop over each dependent variable
for (dep in dependent_damage) {
  #print(dep)
  # Loop over each predictor
  for (pred in selected_predictors) {
    print(pred)
    # Fit the model
   # formula <- as.formula(paste(dep, "~ s(", pred, ", by = f_year, k = 8) + s(year, k=5) +s(x, y, bs = 'gp', k = 20)"))
    formula <- as.formula(paste(dep, "~ s(", pred, ", k = 8)"))
    
    #print(formula)
    model <- gamm(formula,  family = tw,
                  method = 'REML',  
                  data = fin_dat_damage,
                  correlation = corAR1(form = ~ year | pairID))
    
    # Extract model summary
    model_summary <- summary(model$gam)
    
    # Store the AIC value and deviance explained
    model_metrics_damage <- rbind(model_metrics_damage, data.frame(Predictor = pred, 
                                                                 Dependent = dep, 
                                                                 AIC = AIC(model$lme), 
                                                                 R_squared = round(model_summary$r.sq*100,1)
    ))
  }
}

# View the AIC values and deviance explained

(model_metrics_damage)



## Find predicors  lags: RS   -------------
dependent_RS <-  c("RS_wind_beetle")

# Initialize a data frame to store AIC values and deviance explained
model_metrics_RS <- data.frame(Predictor = character(), 
                                   Dependent = character(), 
                                   AIC = numeric(), 
                                   R_squared = numeric())


# Loop over each dependent variable
for (dep in dependent_RS) {
  #print(dep)
  # Loop over each predictor
  for (pred in selected_predictors) {
    print(pred)
    # Fit the model
    #formula <- as.formula(paste(dep, "~ s(", pred, ", by = f_year, k = 8) + s(year, k=4) +s(x, y, bs = 'gp', k = 20)"))
    formula <- as.formula(paste(dep, "~ s(", pred, ", k = 8)"))
    #print(formula)
    model <- gamm(formula,  family = tw,
                  method = 'REML',  
                  data = fin_dat_RS,
                  correlation = corAR1(form = ~ year | pairID))
    
    # Extract model summary
    model_summary <- summary(model$gam)
    
    # Store the AIC value and deviance explained
    model_metrics_RS <- rbind(model_metrics_RS, data.frame(Predictor = pred, 
                                                                   Dependent = dep, 
                                                                   AIC = AIC(model$lme), 
                                                                   R_squared = round(model_summary$r.sq*100,1)
    ))
  }
}

# View the AIC values and deviance explained

(model_metrics_RS)

# merge two tables into one
model_lag_counts_damage_RS <- rbind(model_metrics_RS,
                           model_metrics_damage) %>% 
  distinct()


##  Export as a nice table in word: ------------------------------------------
sjPlot::tab_df(model_lag_counts_damage_RS,
               #col.header = c(as.character(qntils), 'mean'),
               show.rownames = FALSE,
               file="outTable/find_lag_predict_damage_by_counts.doc",
               digits = 1) 


# plot the best predictors and lags: ----------------------------------------

# first order the data:

average_R_squared <- model_lag_counts_damage_RS %>%
  group_by(Predictor) %>%
  summarize(avg_R_squared = mean(R_squared)) %>%
  arrange(desc(avg_R_squared))


# Merge the average R_squared back to the original data
model_lag_counts_damage_RS_avg <- model_lag_counts_damage_RS %>%
  left_join(average_R_squared, by = "Predictor") %>%
  mutate(Predictor = fct_reorder(Predictor, R_squared, .desc = TRUE))



model_lag_counts_damage_RS_avg %>% 
  ggplot(aes(x = Predictor, y = R_squared, col = Dependent)) +
  geom_segment(aes(x = Predictor, 
                   xend = Predictor, 
                   y = 0, 
                   yend = R_squared), color = "grey") +
  geom_point(size = 4) +
  theme_classic2() +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1))



library(stringr)

p_lags_observed_damage <- 
  model_lag_counts_damage_RS %>% 
  dplyr::filter(!str_detect(Predictor, "log")) %>%
  mutate(Predictor = ifelse(str_detect(Predictor, "lag"), 
                            Predictor, 
                            paste0(Predictor, "_lag0"))) %>% 
  dplyr::select(-AIC) %>% 
  pivot_wider(names_from = Dependent, 
              values_from = c(R_squared)) %>% 
  arrange(-damage_vol ) %>% 
 # mutate(Predictor = case_when())
  mutate(Predictor = gsub("sum_ips", "Pop. lev.", Predictor)) %>% 
  mutate(Predictor = gsub("agg_doy", "Agg. timing", Predictor)) %>% 
  mutate(Predictor = gsub("peak_doy", "Peak s. timing", Predictor)) %>% 
  mutate(Predictor = gsub("peak_diff", "Peak s. intens.", Predictor)) %>% 
  mutate(Predictor = as.factor(Predictor)) %>%
  mutate(Predictor = fct_reorder(Predictor, -damage_vol)) %>% 
  ggplot() +
    geom_segment( aes(x=Predictor, xend=Predictor, 
                      y=damage_vol, yend=RS_wind_beetle  ), color="grey",lwd = 1.2) +
    geom_point( aes(x=Predictor, y=RS_wind_beetle, color = "RS_wind_beetle"), size=3 ) +
    geom_point( aes(x=Predictor, y=damage_vol,color = "damage_vol" ),  size=3 )  +
    scale_color_manual(values = c("RS_wind_beetle" = rgb(0.2, 0.7, 0.1, 0.5), 
                                  "damage_vol" = rgb(0.7, 0.2, 0.1, 0.5)),
                       labels = c("RS_wind_beetle" = "Remote sensing\n[# pixels]",
                                  "damage_vol" =  expression("Field observation [m"^3*"]"))) +
    labs(y = expression("Adjusted R"^2*"[%]"),
         x = "",
         color = 'Tree mortality\nsource')+
  theme_bw() +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1),
        legend.position = 'right')
  
windows(6,3.5)
p_lags_observed_damage

ggsave(filename = 'outFigs/observation_vs_predictors_lags.png', 
       plot = p_lags_observed_damage, width = 6, 
       height = 3.5, dpi = 300, bg = 'white')


#### Spearman correlation matrix between predictors -------------------------------------------------------

# Extract the relevant columns from the data
data_selected <- df_catch_RS_damage_pairID[selected_predictors]

cor_matrix <- cor(data_selected, use = "pairwise.complete.obs", method = "spearman")
windows()
corrplot::corrplot(cor_matrix)

# inspect relationships:
pairs(damage_vol ~  sum_ips + sum_ips_lag1 + #sum_ips_lag2 +
        agg_doy    + agg_doy_lag1 + # agg_doy_lag2  +
        peak_doy +  peak_doy_lag1 + #peak_doy_lag2 +
        peak_diff +  peak_diff_lag1, # + #  peak_diff_lag2, 
      data = df_catch_RS_damage_pairID)
# fin_dat_damage

# check for outliers? -------------------------------------------------------------

# filter outliers from dependent and the predictors variables! 
ips_damage_clean_no_outliers <- df_catch_RS_damage_pairID %>% 
  mutate(log_damage_vol = log(damage_vol + 1)) %>% 
  dplyr::filter(log_damage_vol>5,
                log_sum_ips > 8.4) #%>% 
  


# DAMAGE VOLUME model: play with model diagnostics ---------------------------------------
# increase the number of iterations to converge the model
control <- list(niterPQL = 50)

fin_dat_damage_no_extremes <- fin_dat_damage %>% 
  dplyr::filter(damage_vol<150000) %>% 
  dplyr::filter(lag1_damage_vol<150000) 


#### test effect of previuos year damage ----------------------------------------------
m.damage.simpl <- gam(damage_vol ~ 
                       # sum_ips +
                        #lag1_damage_vol +
                            s(sum_ips, k = 3) +
                             s(lag1_damage_vol, k = 3) + # Added term for previous beetle counts
                             s(pairID, bs = 're', k = 5) +
                             s(x, y, bs = 'tp', k = 5) +#,
                            s(f_year, bs = 're')
                      ,
                           data = fin_dat_damage_no_extremes,#  fin_dat_damage, 
                           family = tw,
                      select = TRUE)

m.damage.simpl.lag1 <- gam(damage_vol ~ 
                        # sum_ips +
                        sum_ips_lag1 +
                        #lag1_damage_vol +
                        # s(sum_ips, k = 3) +
                        s(lag1_damage_vol, k = 3) + # Added term for previous beetle counts
                        s(pairID, bs = 're', k = 5) +
                        s(x, y, bs = 'tp', k = 5) +#,
                        s(f_year, bs = 're')
                      ,
                      data = fin_dat_damage_no_extremes,#  fin_dat_damage, 
                      family = tw,
                      select = TRUE)

AIC(m.damage.simpl, m.damage.simpl.lag1)

fin_dat_damage %>% 
  ggplot(aes(x = lag1_damage_vol, y = damage_vol)) +
  geom_point() + 
  geom_smooth()

m <- m.damage.simpl
summary(m)


k.check(m)
predict_data <- ggpredict(m, terms = c('sum_ips'))
plot(predict_data)

predict_data <- ggpredict(m, terms = c('lag1_damage_vol'))
plot(predict_data)



fin.m.damage <- m.damage.simpl

appraise(fin.m.damage)
summary(fin.m.damage)
plot(fin.m.damage, page = 1)
gam.check(m.damage.previous3$gam)
k.check(m.damage.previous3$gam)

# check for autocorrelation
# Extract residuals from the model
residuals <- resid(m.damage.previous3$lme, type = "normalized")

# Plot ACF of the residuals
acf(residuals, main="ACF of Model Residuals")
pacf(residuals, main="ACF of Model Residuals")





# RS model: play with model diagnostics ---------------------------------------

# filetr outliers, as this likely catches teh windthrow and not beetles
fin_dat_RS_clean <- fin_dat_RS %>% 
  dplyr::filter(RS_wind_beetle <1700) %>% 
  dplyr::filter(lag1_RS_wind_beetle <1700)


# PREVIOUS RS --------------------------------------

mRS.simpl <- gam(RS_wind_beetle ~ 
                        #sum_ips +
                        #lag1_damage_vol +
                         s(sum_ips, k = 3) +
                        s(lag1_RS_wind_beetle, k = 3) + # Added term for previous beetle counts
                      s(pairID, bs = 're', k = 5) +
                      s(x, y, bs = 'tp', k = 5) +
                            s(f_year, bs = 're')
                      ,
                      data = fin_dat_RS_clean,#  fin_dat_damage, 
                      family = tw,
                 method = 'REML',  
                      select = TRUE)


mRS.simpl.lag1 <- gam(RS_wind_beetle ~ 
                   # sum_ips_lag1 +
                   #lag1_damage_vol +
                   s(sum_ips_lag1, k = 3) +
                   s(lag1_RS_wind_beetle, k = 3) + # Added term for previous beetle counts
                 s(pairID, bs = 're', k = 5) +
                 s(x, y, bs = 'tp', k = 5) +
                       s(f_year, bs = 're')
                 ,
                 data = fin_dat_RS_clean,#  fin_dat_damage, 
                 family = tw,
                 method = 'REML',  
                 select = TRUE)

AIC(mRS.simpl, mRS.simpl.lag1)

summary(mRS.simpl)
summary(mRS.simpl.lag1)

m<-mRS.simpl

summary(m)
k.check(m)
predict_data <- ggpredict(m, terms = c('sum_ips')) #_lag1
predict_data <- ggpredict(m, terms = c('sum_ips_lag1')) #_lag1

plot(predict_data)

predict_data <- ggpredict(m, terms = c('lag1_RS_wind_beetle'))
plot(predict_data)

# check for autocorrelation
# Extract residuals from the model
residuals <- resid(mRS.previous1$lme, type = "normalized")


# the best model!
fin.m.RS <- mRS.simpl

# export models tables
sjPlot::tab_model(fin.m.RS,     file = "outTable/model_trap_RS.doc")
sjPlot::tab_model(fin.m.damage, file = "outTable/model_trap_damage.doc")



# prepare two plots as output: 
# merge together the predicted data by year, log sum ips for RS, NFI
# export in a single plot
# use log scales 



# Effect plots ----------------------------------------------------
summary(fin.m.damage)
summary(fin.m.RS)


# Spagetti plots for years -----------------------------------------------------

library(scales)
# update data to scatter plot 
fin_dat_damage_plot <- fin_dat_damage %>% 
  dplyr::filter(lag1_damage_vol < 150000) %>% 
  dplyr::filter(damage_vol < 150000) %>% 
  mutate(damage_vol       = damage_vol/1000,
         sum_ips          = sum_ips/1000,
         lag1_damage_vol  = lag1_damage_vol/1000)


# Function to filter data and remove corresponding rows by pairID and year
filter_top_95_percent <- function(data, damage_var = "damage_vol") {
  damage_var_sym <- rlang::sym(damage_var)
  
  # Calculate the 99th percentile threshold for damage_vol
  threshold <- quantile(data[[damage_var]], 0.95, na.rm = TRUE)
  
  # Identify the pairIDs and years to remove
  ids_to_remove <- data %>%
    dplyr::filter(!!damage_var_sym > threshold) %>%
    dplyr::select(pairID, year)
  
  # Filter out all records for these pairIDs and years
  filtered_data <- data %>%
    anti_join(ids_to_remove, by = c("pairID", "year"))
  
  return(filtered_data)
}

# Apply the function to your data
df_full_years_to_spagetti_plot95 <- filter_top_95_percent(df_full_years_to_spagetti_plot)


fin_dat_RS_clean_plot <- fin_dat_RS_clean %>% 
  mutate(RS_wind_beetle_ha      = RS_wind_beetle*0.09,
         lag1_RS_wind_beetle_ha = lag1_RS_wind_beetle*0.09,
         sum_ips_lag1           = sum_ips_lag1/1000,
         sum_ips                = sum_ips/1000)




p_spagett_damage <- plot_data_with_average(df_full_years_to_spagetti_plot, #df_full_years_to_spagetti_plot95, #, 
                                           "damage_vol_k", # "damage_vol_k" 
                                           'my_lab', 
                                           group_var = 'pairID',
                                              my_title = expression("Ground survey [*1000 m"^3*"]"))+ 
  #scale_y_log10() +
  scale_y_log10(labels = label_number()) +  # Adjust y-axis labels to avoid scientific notation
  lims(x = c(2015,2021)) #+
  #coord_cartesian(ylim = c(0, 40000)) #+

(p_spagett_damage)
#range(df_full_years_to_spagetti_plot$damage_vol_k)
#hist(df_full_years_to_spagetti_plot$damage_vol_k)


p_spagett_RS<- plot_data_with_average(df_full_years_to_spagetti_plot, 
                                      "RS_wind_beetle_ha",
                                      'my_lab',  
                                      group_var = 'pairID',
                                       my_title = paste('Remote sensing', '[ha]')) +  
  
                                         scale_y_log10() + 
  lims(x = c(2015,2021))

(p_spagett_RS)

# PLOT: field based damage --------------------------

# Assuming 'model' is your glm.nb model
summary(fin.m.damage)
m <- fin.m.damage
p_val_damage       <- format_p_value_label(summary(m), "s(lag1_damage_vol)")
p_val_ips          <- format_p_value_label(summary(m), "s(sum_ips)")


p0 <- ggpredict(fin.m.damage, terms = "lag1_damage_vol [all]", allow.new.levels = TRUE)
p1 <- ggpredict(fin.m.damage, terms = "sum_ips [all]", allow.new.levels = TRUE)

p0$x <- p0$x/1000 # convert to original values
p0$predicted <- p0$predicted/1000
p0$conf.low <- p0$conf.low /1000
p0$conf.high <- p0$conf.high /1000

p1$x <- p1$x/1000 # convert to original values
p1$predicted <- p1$predicted/1000
p1$conf.low <- p1$conf.low /1000
p1$conf.high <- p1$conf.high /1000


p0.damage <- 
  create_effect_plot(data = p0, 
                     avg_data = fin_dat_damage_plot, 
                     x_col = "lag1_damage_vol", 
                     y_col = "damage_vol", 
                     line_color = "darkgrey", 
                     x_title = expression('Tree mortality lag1' ~ "[" * m^3 * " * 1000]"), 
                     #y_title = paste(lab_popul_level, '*1000'), 
                     #my_title = "Effect of Year on Sum IPS", 
                     x_annotate = 50, 
                     lab_annotate = p_val_damage)

(p0.damage)


# Extract confidence intervals
breakpoint_ci <- confint_breakpoints_damage
lower_ci <- breakpoint_ci[1, "CI(95%).low"]  # Lower bound
upper_ci <- breakpoint_ci[1, "CI(95%).up"]  # Upper bound

# convert logs to counts and divide by 1000 to fit my scale
est_breakpoint <- exp(confint_breakpoints_damage[,"Est."]) /1000      # Estimated breakpoint
lower_ci <- exp(lower_ci)/1000             # Lower CI
upper_ci <- exp(upper_ci)/1000             # Upper CI

print(est_breakpoint)
print(lower_ci)
print(upper_ci)


p1.damage <- 
  create_effect_plot(data = p1, 
                     avg_data = fin_dat_damage_plot, 
                     x_col = "sum_ips", 
                     y_col = "damage_vol", 
                     line_color = "#A50026", 
                     x_title = 'Population level [#*1000]' , 
                     y_title = paste(lab_popul_level, '*1000'), 
                     #my_title = "Effect of Year on Sum IPS", 
                     x_annotate = 50, 
                     lab_annotate = p_val_ips) +
  geom_vline(xintercept = est_breakpoint, linetype = "dashed", color = "black") +
   # Add grey shaded rectangle for CI
  annotate("rect", xmin = lower_ci, xmax = upper_ci, ymin = -Inf, ymax = Inf, 
           alpha = 0.2, fill = "grey")

(p1.damage)

##### PLOT RS based damage -----------------------------------------

# Assuming 'model' is your glm.nb model
summary(fin.m.RS)

m <- fin.m.RS
p_val_RS       <- format_p_value_label(summary(m), "s(lag1_RS_wind_beetle)")
p_val_ips          <- format_p_value_label(summary(m), "s(sum_ips)")

p0 <- ggpredict(fin.m.RS, terms = "lag1_RS_wind_beetle [all]", allow.new.levels = TRUE)
p1 <- ggpredict(fin.m.RS, terms = "sum_ips [all]", allow.new.levels = TRUE)

p0$x <- p0$x*0.09 # convert piel counts to hectares
p0$predicted <- p0$predicted*0.09 #/100
p0$conf.low <- p0$conf.low*0.09 #/100
p0$conf.high <- p0$conf.high*0.09# /100

p1$x <- p1$x/1000 # convert to original values
p1$predicted <- p1$predicted*0.09
p1$conf.low <- p1$conf.low *0.09
p1$conf.high <- p1$conf.high *0.09


p0.RS <- 
  create_effect_plot(data = p0, 
                     avg_data = fin_dat_RS_clean_plot, 
                     x_col = "lag1_RS_wind_beetle_ha", 
                     y_col = "RS_wind_beetle_ha", 
                     line_color = "darkgrey", 
                     x_title = 'Tree mortality lag1 [ha]' , 
                     #y_title = paste(lab_popul_level, '*1000'), 
                     #my_title = "Effect of Year on Sum IPS", 
                     x_annotate = 75, 
                     lab_annotate = p_val_RS)

(p0.RS)

# Extract confidence intervals
breakpoint_ci <- confint_breakpoints_RS
lower_ci <- breakpoint_ci[1, "CI(95%).low"]  # Lower bound
upper_ci <- breakpoint_ci[1, "CI(95%).up"]  # Upper bound

# convert logs to counts
est_breakpoint <- exp(breakpoint_ci[,"Est."]) /1000      # Estimated breakpoint
lower_ci <- exp(lower_ci)/1000             # Lower CI
upper_ci <- exp(upper_ci)/1000             # Upper CI



p1.RS <- 
  create_effect_plot(data = p1, 
                     avg_data = fin_dat_RS_clean_plot, 
                     x_col = "sum_ips", 
                     y_col = "RS_wind_beetle_ha", 
                     line_color = "#A50026", 
                     x_title = 'Population level [#*1000]' , 
                     #y_title = paste(lab_popul_level, '*1000'), 
                     #my_title = "Effect of Year on Sum IPS", 
                     x_annotate = 40, 
                     lab_annotate = p_val_ips) +
  geom_vline(xintercept = est_breakpoint, linetype = "dashed", color = "black") +
  # Add grey shaded rectangle for CI
  annotate("rect", xmin = lower_ci, xmax = upper_ci, ymin = -Inf, ymax = Inf, 
           alpha = 0.2, fill = "grey")


(p1.RS)

p.comp2 <- ggarrange(
  p_spagett_damage, 
  #p0.damage, 
  p1.damage, 
  p_spagett_RS, 
  #p0.RS, 
  p1.RS, 
  ncol = 2, nrow = 2, align = 'hv',
  labels = c('[a]', '[b]', '[c]', '[d]'),# , '[e]', '[f]'
  font.label = list(size = 10, face = "plain"),  # Set font to normal (plain) and specify size
  label.x = 0.15,  # Adjust these values to position the labels within the plot
  label.y = 0.85
)
(p.comp2)


ggsave(filename = 'outFigs/observation_vs_traps2.png', 
       plot = p.comp2, width = 4, height = 4, dpi = 300, bg = 'white')


##### quick plotting -----------------------------------------------

# create area plots: sum damage per year
df_sum <- df_merged_RS_damage_district %>% 
  group_by(year) %>%
  dplyr::summarise(#RS_harvest = sum(RS_harvest, na.rm = T),
                   RS_wind_beetle  = sum(RS_wind_beetle , na.rm = T),
                   damaged_volume_total_m3  = sum(damaged_volume_total_m3,na.rm = T ))
  

p1 <- df_sum %>% 
  ggplot(aes(x = year,
             y = RS_wind_beetle*0.09)) +
  geom_line(col = 'green', lty = 'dashed', lwd = 1.5) + 
  labs(y = 'RS tree mortality [ha]') + 
  theme_classic()


p2<- df_sum %>% 
  ggplot(aes(x = year,
             y = damaged_volume_total_m3/10000)) +
  geom_line(col = 'red', lwd = 1.5) +
  labs(y = 'Reported tree damage [*10000 m^3]') + 
  theme_classic()

# df_sum
windows(7,3)
ggarrange(p1, p2, nrow = 1, ncol = 2, align = 'hv', labels = c('[a]', '[b]'))



# MAPS -----------------------------------------------------------------------------------

## Correlate between RS and volume data m3 ---------------------------------------------------
df_cor <- df_merged_RS_damage_district %>% 
  dplyr::filter(!ID %in% c(0, 50104, 60613,62705)) %>% # exlude if there is missing data
  group_by(ID) %>% 
  dplyr::summarize(spearm_cor_beetle = cor(damaged_volume_total_m3, RS_wind_beetle, 
                                           method = "spearman", use = "complete.obs"))


summary(df_cor)

df_merged_RS_damage_district %>% 
  dplyr::filter(RS_wind_beetle < 10000) %>% 
  ggplot(aes(x = RS_wind_beetle, y = damaged_volume_total_m3)) + 
  geom_point() + 
  geom_smooth(method = "lm") + 
  stat_cor(method = "pearson", label.x = 3000, label.y = max(df_merged_RS_damage_district$damaged_volume_total_m3, na.rm = TRUE))


cor_shp <- sf_simpled %>% 
  left_join(df_cor, by = c("forstrev_1" = "ID"))

all_shp <- sf_simpled %>% 
  left_join(df_merged_RS_damage_district, by = c("forstrev_1" = "ID"))

df_damage_sum <- df_merged_RS_damage_district %>% 
  group_by(ID) %>% 
  summarise(sum_dmg  =sum(damaged_volume_total_m3, na.rm = T),
            sum_RS  =sum(RS_wind_beetle, na.rm = T)) %>% 
  dplyr::filter(ID!=0) %>% 
  ungroup(.)

# add data to the full geometry to run teh intersection with the trap data
#all_shp_complex <- sf_districts_trans %>% 
#  left_join(df_merged_RS_damage_district, by = c("forstrev_1" = "ID"))

dmg_shp <- 
  sf_simpled %>% 
    dplyr::select(c(forstrev_1)) %>% 
  full_join(df_damage_sum, by = c("forstrev_1" = "ID")) %>% 
  dplyr::select(c(forstrev_1, sum_dmg,sum_RS)) #%>% 
  #mutate(log_sum_dmg )
  
  

# cap values at 2e05

hist(dmg_shp$sum_dmg)
hist(dmg_shp$sum_RS)

 

 # Calculate the XX percentile of the sum_dmg column
 perc95_dmg <- quantile(dmg_shp$sum_dmg, 0.95)
 perc95_RS <- quantile(dmg_shp$sum_RS, 0.95)
 


# cap values at 95% interval
dmg_shp <- dmg_shp %>% 
  mutate(dmg_cap = case_when(sum_dmg > perc95_dmg~perc95_dmg, TRUE ~ sum_dmg),
         RS_cap = case_when(sum_RS > perc95_RS ~ perc95_RS, TRUE ~ sum_RS))

 #sum coundt of beetles by locations 

df_sum_ips_years <- sum_ips %>% 
   mutate(pairID = str_sub(falsto_name, 1, -3)) %>% 
  group_by(pairID) %>% 
  summarise(sum = sum(sum_ips, na.rm = T)/2)  # divide by two to have average sum per pair

 # merge with simpler trap pairs (sites)
 df_sum_ips_years_sf <- xy_sf_fin %>% 
   mutate(pairID = str_sub(falsto_name, 1, -3)) %>% 
   group_by(pairID) %>% 
   slice(1) %>%  # Keep only the on trap per pair
   ungroup(.) %>% 
   full_join(df_sum_ips_years,by = join_by(pairID)) %>% 
   dplyr::select(pairID, sum) #%>% 
   
 
 # Calculate the XX percentile of the sum_dmg column
 perc95_sum_ips <- quantile(df_sum_ips_years_sf$sum, 0.95)
 
 df_sum_ips_years_sf <- df_sum_ips_years_sf %>% 
   mutate(sum_cap = case_when(sum > perc95_sum_ips~perc95_sum_ips, TRUE ~ sum))
 
 
 hist(df_sum_ips_years_sf$sum)
 
 bav_shp <- st_read('rawData/outline_bavaria.gpkg')
 bav_3035 <- st_transform(bav_shp, crs(dmg_shp))
 
 #4298150 
 
 
# ad damage data to district geometry
## MAP: plot correlation with beetle counts & tree damage ----------------------------------
# Extract colors from RColorBrewer's YlOrRd palette
colors <- rev(brewer.pal(9,  "RdYlBu" )) # "YlOrRd"  "RdYlBu"


map.damage <- 
  ggplot(dmg_shp) +
   geom_sf(data = dmg_shp,
          aes(fill = dmg_cap/1000)) +
    geom_sf(data = bav_3035,
            color = 'black', fill = NA, lwd = 0.8)  +
    geom_sf(data = df_sum_ips_years_sf,
            aes(color = sum_cap/1000), color = 'grey20', size = 2.5, stroke = 1) +
    geom_sf(data = df_sum_ips_years_sf,
            aes(color = sum_cap/1000), size = 2) +
  scale_color_gradientn(
    name = expression("Beetle catch [#*1000]"^" "),
  #  name = expression("Beetle catch [1000 m"^" "*"]"),
  #  name = expression("Beetle catch [1000 m"^3*"]"),  # working
   # name = expression("Beetle catch [#*1000]"^" "), 
   # name = "Beetle catch [#*1000]",
    colors = colors,
    guide = guide_colorbar(title.position = "top", title.hjust = 0.5)
  )  +
    scale_fill_gradientn(
      name = expression("Tree mortality [1000 m"^3*"]"), 
      #name = bquote("Tree mortality [1000 m"^3*"]"), 
    #name ="Tree mortality [1000 m3]", # bquote(Tree~mortality~('1000'~m^3)), #expression(Tree~mortality~('1000'm^3)) , #"Tree mortality [m3]",
    colors = colors,
    guide = guide_colorbar(title.position = "top", title.hjust = 0.5)
  )  +
  labs(title = "[a]") +
  theme_void() +
  theme(
    legend.position = "bottom",
    legend.title = element_text(angle = 0, size = 8),  # Smaller title
    legend.text = element_text(size =8),              # Smaller text
    legend.box = "horizontal",  
     legend.key.width = unit(0.5, "cm"),                  # Adjust width
    legend.key.height = unit(0.2, "cm"),               # Adjust height
    legend.justification = "center",  # Ensures the legend is fully centered
    plot.title = element_text(hjust = 0, size = 8 )             # Center title
  )

  #annotation_scale(location = "bl", width_hint = 0.5) +  # Add scale bar at bottom left
  #annotation_north_arrow(location = "tl", which_north = "true", style = north_arrow_fancy_orienteering)  # Add north arrow at top left
map.damage








map.RS <- 
  ggplot(dmg_shp) +
   geom_sf(data = dmg_shp,
          aes(fill = RS_cap*0.09)) +
  geom_sf(data = bav_3035,
          color = 'black', fill = NA, lwd = 0.8)  +
  geom_sf(data = df_sum_ips_years_sf,
          aes(color = sum_cap/1000), color = 'grey20', size = 2.5, stroke = 1) +
  geom_sf(data = df_sum_ips_years_sf,
          aes(color = sum_cap/1000), size = 2) +
  scale_color_gradientn(
    name = "Beetle catch [#*1000]",
    colors = colors,
    guide = guide_colorbar(title.position = "top", title.hjust = 0.5)
  )  +
  scale_fill_gradientn(
    name = "Tree mortality [ha]",
    colors = colors,
    guide = guide_colorbar(title.position = "top", title.hjust = 0.5)
  )  +
  labs(title = "[b]") +
  theme_void() +
  theme(
    legend.position = "bottom",
    legend.title = element_text(angle = 0, size = 8),  # Smaller title
    legend.text = element_text(size = 8),              # Smaller text
    legend.key.width = unit(0.5, "cm"),                  # Adjust width
    legend.key.height = unit(0.2, "cm") ,               # Adjust height
    plot.title = element_text(hjust = 0, size = 8)             # Center title
  )

windows(7,5)
damage_maps <- ggarrange(map.damage, map.RS, align = 'hv')
damage_maps

ggsave(filename = 'outFigs/damage_maps_R1.png', 
       plot = damage_maps, width = 7, 
       height = 5, dpi = 300, bg = 'white')


#p_correlograms_ribbons
library(svglite)

ggsave(filename = 'outFigs/damage_maps.svg', plot = damage_maps,
       width = 7, height = 5, dpi = 300,
       bg = 'white',
       device = "svg")




# MAP: plot correlation with beetle counts & tree damage ----------------------------------
ggplot(cor_shp) +
   geom_sf(color = 'grey93', 
            fill  = 'grey93') + 
  geom_sf(data = cor_shp_counts_dmg,
          aes(fill = value)) +
  scale_fill_gradient2(low = "blue", mid = "white", high = "red", 
                       midpoint = 0,  # Set the midpoint at 0 for white color
                       limit = c(min(cor_shp_counts_dmg$value, na.rm = TRUE), 
                                 max(cor_shp_counts_dmg$value, na.rm = TRUE)),
                       name = "Spearman\nCorrelation") +
  labs(title = "Spearman Correlation Coefficient",
       subtitle = "between spruce damage [m3] and beetle sum/year by region") +
  theme_void()  + 
  facet_wrap(.~name)#





# Plots & maps ------------------------------------------------------------


### RS plots -----------------------------------------------------------------


#### show spearman correlation coefficient  -------
# between volume damage and RS observation (wind, beetle)

# wind_beetle vs damage volume 

p1 <- ggplot(cor_shp) +
  # geom_sf(color = 'black', 
  #          fill  = 'grey93') #+ 
  geom_sf(data = cor_shp,
          aes(fill = spearm_cor_beetle)) +
  scale_fill_gradient2(low = "blue", mid = "white", high = "red", 
                       midpoint = 0,  # Set the midpoint at 0 for white color
                       limit = c(min(cor_shp$spearm_cor_beetle, na.rm = TRUE), 
                                 max(cor_shp$spearm_cor_beetle, na.rm = TRUE)),
                       name = "Spearman\nCorrelation") +
  labs(title = "Spearman Correlation Coefficient",
       subtitle = "Correlation between ground survey data and\nRS based estimation [ha] by forest district") +
  theme_minimal()  #

p1
summary(cor_shp$spearm_cor_beetle)
mean(cor_shp$spearm_cor_beetle, na.rm = T)
sd(cor_shp$spearm_cor_beetle, na.rm = T)
median(cor_shp$spearm_cor_beetle, na.rm = T)


# Min. 1st Qu.  Median    Mean 3rd Qu.    Max.    NA's 
# -0.9856  0.1518  0.4928  0.4412  0.8117  1.0000       3 





### Map damage [m3] per year ------------------------------------------------

summary(all_shp$damaged_volume_total_m3)

# Define the 75th percentile
damage_75 <- quantile(all_shp$damaged_volume_total_m3, probs = 0.75)
RS_wind_beetle_75 <- quantile(all_shp$RS_wind_beetle , probs = 0.75, na.rm = T)

# cap tthe values for better visualization of data
all_shp <- all_shp %>% 
  mutate(damaged_volume_cap = case_when(damaged_volume_total_m3 >= damage_75 ~ damage_75,
                                        TRUE ~ damaged_volume_total_m3),
         RS_wind_beetle_cap = case_when(RS_wind_beetle >= RS_wind_beetle_75 ~ RS_wind_beetle_75,
                                        TRUE ~ RS_wind_beetle))


# Plot the map
p1<-ggplot(all_shp) +
  geom_sf(aes(fill = damaged_volume_cap)) +
  scale_fill_gradient(low = "yellow", high = "red", 
                      name = expression("Tree damage [m"^3*"]"),  
                      na.value = 'transparent') +
  facet_wrap(. ~ year) +
  theme(legend.position = 'bottom') +
  theme_void() 


p2<-ggplot(all_shp) +
  geom_sf(aes(fill = RS_wind_beetle_cap*0.09)) +
  scale_fill_gradient(low = "yellow", high = "red", 
                      name = expression("RS damage [ha]"),  
                      na.value = 'transparent') +
  facet_wrap(. ~ year) +
  theme(legend.position = 'bottom') +
  theme_void() 


ggarrange(p1, p2, labels = c('[a] Damaged volume', '[b] RS Damaged area'))

# plot gam between damage and RS beetle ------------------------------------------
df_merged_RS_damage_district %>% 
  ggplot(aes(x = log(RS_wind_beetle +1) ,
             y = log(damaged_volume_total_m3) )) +
  geom_point() +
  geom_smooth() + 
  facet_wrap(.~year, scales = 'free')


# Summary tables ---------------------------------

## for year & per district --------------------------------------

# Represent results using quantiles, as they are skewed?
#qntils = c(0, 0.25, 0.5, 0.75, 1)
df_merged_RS_damage_district %>% 
  ungroup() %>% 
  summarize(sum_volume = sum(damaged_volume_total_m3),
            sum_RS = sum(RS_wind_beetle, na.rm = T)*0.09)

## data per years
observation_mortality_year <- 
  df_merged_RS_damage_district %>% 
  ungroup(.) %>% 
  filter(year %in% 2015:2021) %>%
  dplyr::select(ID, year, damaged_volume_total_m3, RS_wind_beetle) %>% 
  dplyr::mutate(RS_wind_beetle = RS_wind_beetle*0.09) %>% # convert pixels to hectares
  group_by(year) %>% 
  dplyr::summarize(mean_dam   = mean(damaged_volume_total_m3  , na.rm = T),
                   sd_dam     = sd(damaged_volume_total_m3, na.rm = T),
                   mean_RS   = mean(RS_wind_beetle  , na.rm = T),
                   sd_RS     = sd(RS_wind_beetle, na.rm = T)) %>% 
  mutate(Field_volume       = stringr::str_glue("{round(mean_dam,1)}±{round(sd_dam,1)}"),
         RS_area            = stringr::str_glue("{round(mean_RS,1)}±{round(sd_RS,1)}")) %>% 
  dplyr::select(year, Field_volume, RS_area) 


(observation_mortality_year)

## per drought years 
observation_mortality_drought <- 
  df_merged_RS_damage_district %>% 
  ungroup(.) %>% 
  filter(year %in% 2015:2021) %>%
  mutate(drought_status = ifelse(year %in% 2018:2020, "Hotter drought", "Other")) %>% 
  dplyr::select(ID, drought_status, damaged_volume_total_m3, RS_wind_beetle) %>% 
  dplyr::mutate(RS_wind_beetle = RS_wind_beetle*0.09) %>% # convert pixels to hectares
  group_by(drought_status) %>% 
  dplyr::summarize(mean_dam   = mean(damaged_volume_total_m3  , na.rm = T),
                   sd_dam     = sd(damaged_volume_total_m3, na.rm = T),
                   mean_RS   = mean(RS_wind_beetle  , na.rm = T),
                   sd_RS     = sd(RS_wind_beetle, na.rm = T)) %>% 
  mutate(Field_volume       = stringr::str_glue("{round(mean_dam,1)}±{round(sd_dam,1)}"),
         RS_area            = stringr::str_glue("{round(mean_RS,1)}±{round(sd_RS,1)}")) %>% 
  dplyr::select(drought_status, Field_volume, RS_area) 


(observation_mortality_drought)


## Export as a nice table in word: ----------------------------------

sjPlot::tab_df(observation_mortality_year,
               #               col.header = c(as.character(qntils), 'mean'),
               show.rownames = F,
               file="outTable/summary_out_observation_damage_year.doc",
               digits = 0) 

sjPlot::tab_df(observation_mortality_drought,
               show.rownames = F,
               file="outTable/summary_out_observation_mortality_drought.doc",
               digits = 0)
